+++ app/src/main/java/com/example/projeto001/fragments/PrimeiroFragment.kt
@@ -1,0 +1,18 @@
+package com.example.projeto001
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+class PrimeiroFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_primeiro, container, false)
+    }
+}
