+++ app/src/main/java/com/example/projeto001/model/Geolocation.kt
@@ -1,0 +1,3 @@
+package com.example.projeto001.model
+
+data class Geolocation()
