+++ app/src/main/java/com/example/projeto001/service/ServiceUser.kt
@@ -1,0 +1,4 @@
+package com.example.projeto001.service
+
+interface ServiceUser {
+}
