+++ app/src/main/java/com/example/projeto001/ui/MainActivity.kt
@@ -2,10 +2,17 @@ package com.example.projeto001
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+    }
+
+    fun replaceFragment(fragment: Fragment){
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.container1, fragment)
+            .commitNow()
     }
 }
