+++ app/src/main/java/com/example/projeto001/ui/MainActivity2.kt
@@ -1,0 +1,11 @@
+package com.example.projeto001.ui
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class MainActivity2 : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main2)
+    }
+}
